+++ tests/db_config.py
@@ -61,8 +61,6 @@ MIMERPY_STABLE = os.environ.get('MIMERPY_STABLE', 'True')
 MIMERPY_TRACE = os.environ.get('MIMERPY_TRACE')
 
 def setup():
-    if MIMERPY_TRACE:
-        mimerpy._trace()
     syscon = mimerpy.connect(**SYSUSR)
     with syscon.cursor() as c:
         try: